pyproject.toml
[tool.poetry]
name = "draw_stream"
version = "1.0.0"
description = "Interactive donation-driven pixel art pipeline"
authors = ["Draw Stream Team <team@example.com>"]
readme = "README.md"
packages = [{ include = "draw_stream", from = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.119.1"
httpx = "^0.28.1"
orjson = "^3.11.3"
pydantic = "^2.12.3"
pydantic-settings = "^2.11.0"
pygame = "^2.6.1"
python-dotenv = "^1.1.1"
uvicorn = { version = "^0.38.0", extras = ["standard"] }
websockets = "^15.0.1"
Pillow = "^10.4.0"
numpy = "^2.1.2"
scipy = "^1.13.0"
"json-repair" = "^0.52.2"
diffusers = "^0.29.0"
transformers = "^4.44.2"
safetensors = "^0.4.4"
scikit-image = "^0.24.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
pytest-cov = "^5.0.0"
mypy = "^1.13.0"
ruff = "^0.6.9"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "strict"
pythonpath = ["src"]
testpaths = ["tests"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
